Refactor/(Changed userauth to userStore with pinia)
