+++ frontend/src/main.js
@@ -3,13 +3,13 @@ import './assets/main.css'
 import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router/router'
-import {userAuth} from './composables/userAuth.js'
+import { createPinia } from 'pinia'
 
 const app = createApp(App)
 
-const auth = userAuth()
-auth.checkAuth()
+const pinia = createPinia()
 
+app.use(pinia)
 app.use(router)
 
 app.mount('#app')
