+++ frontend/src/router/router.js
@@ -1,9 +1,10 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
 import LogInView from '../views/LogInView.vue'
-import { userAuth } from '../composables/userAuth.js'
 import RegisterView from '../views/RegisterView.vue'
 import ProfileView from '../users/ProfileView.vue'
+import { useUserStore } from '@/stores/userStore'  
+import { getActivePinia } from 'pinia'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -26,21 +27,24 @@ const router = createRouter({
     {
       path: '/register',
       name: 'register',
-      component: RegisterView,    
+      component: RegisterView,
     },
     {
       path: '/profile',
       name: 'profile',
       component: ProfileView,
+      meta: { requiresAuth: false }, //endre til true senere
     },
   ],
 })
 
 router.beforeEach((to, from, next) => {
-  const {isAuthenticated, checkAuth} = userAuth()
-  checkAuth()
+  const pinia = getActivePinia()
+  const userStore = useUserStore(pinia)
 
-  if(to.meta.requiresAuth && !isAuthenticated.value) {
+  userStore.checkAuth()
+
+  if (to.meta.requiresAuth && !userStore.isAuthenticated) {
     next('/login')
   } else {
     next()
