+++ frontend/src/styles/LogInView.css
@@ -2,13 +2,14 @@
   max-width: 400px;
   margin: 4rem auto;
   padding: 2rem;
-  background: #ffffff;
 }
 
 .logInPage h1 {
   text-align: center;
   margin-bottom: 2rem;
-  font-size: 2rem;
+  font-size: 2.4rem;
+  font-weight: bold;
+  font-family: 'Montagu Slab', serif;
   color: #333;
 }
 
@@ -16,6 +17,8 @@
   display: block;
   margin: 1rem 0 0.25rem;
   font-weight: 500;
+  font-family: 'Abel', sans-serif;
+  font-size: 1rem;
 }
 
 .logInWrapper {
